Clean up repo - reorganize
* Move the models into a subdir.
* Move the utils into the `utils` subdir.
* Move the catalog-related files into the `catalog` subdir.
* Add __init__ files for subdirs.
